requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP client for Slack API calls
httpx>=0.25.2

# AI/ML dependencies
cohere>=4.37.0
numpy>=1.26.0

# Data validation
pydantic>=2.5.0

# Development dependencies (optional)
python-dotenv>=1.0.0

# Monitoring and metrics
prometheus-client>=0.19.0
structlog>=23.2.0

# Performance and caching
redis>=5.0.0

# Phase 5: Production Database Integration
# Vector database
lancedb>=0.3.0
pyarrow>=12.0.0

# Relational database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations

# Data validation
pydantic>=2.5.0
